SyntaxAnalyzer: Check path with extension() instead of ends_with()
ends_with() only matches whole pieces of file paths (parts in between
backslashes) which means checking if 'path/to/test.jack'
ends_with(".jack") will always return false because ends_with can only
see 'path', 'to', and 'test.jack'.

extension() returns the extension of the file at the end of the path so
we can then compare that with "jack".