Lexer: Implement lex_comment and its required functions
lex_comment() gets called when the current token is '/'. It checks that
the next token is also a '/'. If it is, it advances our index to the
next line, ignoring the comment. Otherwise, it returns an error telling
the user that they have a floating '/' that makes no sense and hints to
them they probably meant '//'.