SyntaxAnalyzer: Write parse tree out to xml files
If given a directory, xml files are written back into that directory.
One xml file per jack file in the directory and starting with the name
of the jack file (ex: test.jack -> test.xml).

If given a file, xml file is written to the current directory the user
is in. If given "test.jack" it will write out to "./test.xml".