Lexer: Don't treat all backslashes as comments - division exists :)
By treating all backslashes as either the start or end of a comment,
division would throw the single backslash comment error which wasn't
what was happening at all.

If the current token is a backslash peek at the next token to see if it
is a backslash or an asterik, if it is treat it as a comment. In
treating it as a comment, we should advance all the way past the closing
backslash so that it does not get tried again as a token. If the peek is
neither a backslash or an asterik, push the backslash token into the
parse tree.