Lexer: Specify SyntaxAnalyzer output path in DEBUG mode
I don't want to have to crawling through my file tree because I
wrote the output to some dir on the total opposite side. If not
debugging, then yes it should output to the path given to it, but for
debugging it's annoying.

Eventually, there should be a -o flag to specify path and output name.