Parser: Implement overarching parse_statements function
parse_statements if the jump off point for parsing all the statements in
a subroutine body. This function looks at the current token which should
be the start of a statement or the end of a subroutine and calls the
appropriate functions. It then calls itself recursively until there are
no more statements left to parse in the subroutine body. At which point
it should leave the index on the closing curly and return its parse tree
to the caller.