Parser: Correct bugs in parse_expression_list
For one, actually return the expression list parse tree instead of an
empty string... haha. Also, call parse_expression first, then check if
there is a comma, and call parse_expression again if there is. Before
calling the first time, check to make sure that it is not an empty
expression list such as 'do draw();'.