Parser+SymbolTable: Revert returning (String, String) + focus on
getting the symbol table right. We want to make sure our symbol table is
right before diving into codegen. To do we have to gather information
from the parser for each identifier and add that identifier to our
symbol table.

Rewrite parse_multi_variable_declaration to take in a ty and kind. This
is because the type and kind don't change for each variable in the
multi variable declaration. They are all of one type and kind. Consider:
field int x, y, z. x y and z are all of type int and kind field.